confirmation added

+++ app/Http/Controllers/BusRoutesController.php
@@ -63,11 +63,16 @@ class BusRoutesController extends Controller
 
         //Create BusRoute
         $BusRoute = new BusRoute;
-        $BusRoute->title = $request->input('routename');
+        $BusRoute->routename = $request->input('routename');
         $BusRoute->user_id = Admin::user()->id;
-        
         $BusRoute->save();
-        return redirect('/admin/auth/routes/create')->with('success','Bus Route Created Successfully!');
+        $success = array (
+            'title' => 'Add Route for Bus',
+            'message' => 'Bus Route Created Successfully!'
+        );
+        $name = $BusRoute->routename;
+        return redirect('/admin/auth/routes/create')->with('success','"'.$name.'" => Bus Route Created Successfully!');
+        //return redirect('/admin/auth/routes/create')->with('success');
     }
 
     /**
@@ -79,7 +84,7 @@ class BusRoutesController extends Controller
     public function show($id)
     {
         $BusRoute = BusRoute::find($id);
-        return view('BusRoutes.show')->with('BusRoute',$BusRoute);
+        return view('busroutes.show')->with('BusRoute',$BusRoute);
     }
 
     /**
@@ -93,7 +98,7 @@ class BusRoutesController extends Controller
         $title = 'Edit Bus Route';
         $titleInfo = 'Available Bus Routes';
         $BusRoute = BusRoute::find($id);
-        $BusRoutes =  BusRoute::orderBy('routename')->paginate(20);
+        $BusRoutes = BusRoute::orderBy('routename')->paginate(20);
         $data = array(
             'title' => $title,
             'BusRoutes' => $BusRoutes,
@@ -105,7 +110,7 @@ class BusRoutesController extends Controller
         {
             return redirect('/admin/auth/routes/create')->with('error','Unauthorized Access Denied!');
         } */
-        return view('BusRoutes.edit')->with($data); 
+        return view('busroutes.edit')->with($data); 
 
     }
 
@@ -130,12 +135,14 @@ class BusRoutesController extends Controller
         ]);
 
         $BusRoute = BusRoute::find($id);
-        $BusRoute->title = $request->input('routename');
+        $name = $BusRoute->routename;
+        $BusRoute->routename = $request->input('routename');
+        $name2 = $BusRoute->routename;
         if(Admin::user()->id !== $BusRoute->user_id){
             $BusRoute->user_id = Admin::user()->id;
         }
         $BusRoute->save();
-        return redirect('/admin/auth/routes/create/'.$id)->with('success','Bus Route Updated Successfully!');
+        return redirect('/admin/auth/routes/create/')->with('success','"'.$name.'" => Updated to "'.$name2.'". Bus Route Updated Successfully!');
     }
 
     /**
@@ -159,8 +166,20 @@ class BusRoutesController extends Controller
             //Delete Image From Windows Directory
             Storage::delete('public/cover_images/'.$BusRoute->cover_image);
         } */
-
-        $BusRoute->delete();
-        return redirect('/admin/auth/routes/create/')->with('success','Bus Route Removed Successfully!');
+        $points = DB::table('points')->where('routeid', $id)->first();
+        $name = $BusRoute->routename;
+        if($points){
+            if($points->routeid == $id){
+                return redirect('/admin/auth/routes/create/')->with('error', '"'.$name.'" => This Bus Route has assigned one or more bus stop point. Delete all bus stop point related to the route. Then delete the route.');
+            }
+            else{
+                return redirect('/admin/auth/routes/create/')->with('error', '"'.$name.'" => Something went worng!!');
+            
+            }
+        }else{
+            $BusRoute->delete();
+            return redirect('/admin/auth/routes/create/')->with('success','"'.$name.'" => Bus Route Removed Successfully!');
+        }
+        
     }
 }
